requirements.txt
mando
oauth2client
google-api-python-client
google-genomics-pipeline-runner